+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Check detailed migration documentation for breaking changes in [7.0.0](./doc/MIGRATION-v7.md)
 * `Config.errorMessages` has been removed. Please use Kubernetes status messages directly.
 * Fix #6138: Removed unused `io:fabric8:kubernetes-model` artifact
+* Fix #6156: Removed deprecated extension `io:fabric8:service-catalog`
 
 ### 6.13.1 (2024-07-02)